use simpler Pixi.js APIs for canvas renderer
In non WebGL browsers Pixi.js-legacy handles advanced render methods poorly
