+++ src/layers/HoleSizeLayer.ts
@@ -3,7 +3,7 @@ import { HoleSizeLayerOptions, OnMountEvent, OnUpdateEvent, OnRescaleEvent, Hole
 import { makeTubularPolygon } from '../datautils/wellboreItemShapeGenerator';
 import { createNormals, offsetPoints } from '../utils/vectorUtils';
 import { HOLE_OUTLINE } from '../constants';
-import { Point } from 'pixi.js';
+import { Point, RENDERER_TYPE } from 'pixi.js';
 
 export class HoleSizeLayer extends WellboreBaseComponentLayer {
   options: HoleSizeLayerOptions;
@@ -48,7 +48,7 @@ export class HoleSizeLayer extends WellboreBaseComponentLayer {
       return;
     }
 
-    const { maxTextureDiameterScale } = this.options;
+    const { maxTextureDiameterScale, firstColor } = this.options;
 
     const texture = this.createTexture(holeObject.diameter * maxTextureDiameterScale);
 
@@ -66,11 +66,14 @@ export class HoleSizeLayer extends WellboreBaseComponentLayer {
     }
 
     const polygonCoords = makeTubularPolygon(leftPath, rightPath);
-
-    this.drawRope(
-      pathPoints.map((p) => new Point(p[0], p[1])),
-      texture,
-    );
+    if (this.renderType === RENDERER_TYPE.CANVAS) {
+      this.drawBigPolygon(polygonCoords, firstColor);
+    } else {
+      this.drawRope(
+        pathPoints.map((p) => new Point(p[0], p[1])),
+        texture,
+      );
+    }
 
     this.drawLine(polygonCoords, lineColor, HOLE_OUTLINE, true);
   };
